fix casting problem ex00

+++ cpp00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: maolivei <maolivei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/13 20:39:11 by maolivei          #+#    #+#             */
-/*   Updated: 2022/11/13 20:39:16 by maolivei         ###   ########.fr       */
+/*   Updated: 2022/11/15 20:47:34 by maolivei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int main(int argc, char **argv)
         std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
         return (0);
     }
-    for (size_t i = 1; i < argc; ++i)
+    for (size_t i = 1; i < (size_t)argc; ++i)
         str.append(argv[i]);
     for (std::string::iterator c = str.begin(); c != str.end(); ++c)
         *c = toupper(*c);
